Speech-Text-Analysis/requirements.txt
streamlit==1.25.0
transformers==4.33.3
torch==2.1.0
nltk==3.8.1
librosa==0.10.0
soundfile==0.12.1
numpy==1.24.3
openai-whisper
audio-recorder-streamlit
numba==0.61.0
tbb==2021.9.0
